fix(log): [preview] reset log table state on pagination change
Not resetting table state caused filters to persist from previous page
which in turn caused some filters to be missing from default like:
if we didn't have `MI` level on the first page and user navigated to
second log page `MI` level appears in the filters for level but is not
applied by default. Fix this by resetting whole table state on
navigation.

Signed-off-by: Danil Kostromin <danil.kostromin@icloud.com>